chore(clean-code): split oversized components + dedupe shared UI (#106)
Pure refactor, no behaviour change (typecheck + lint + 218 tests green, visually
identical on Overview / Faction).

- Shared: lib/rankMedal (medal colours), components/ButtonLink (Primary/Secondary),
  FactionStandingsBar (the faction bars used in 2 places).
- Leaderboard: extract LeaderboardRow into its own file; drop the duplicated
  rankColorClass.
- Overview: extract OverviewHeroCard + OverviewFactionCard; use shared ButtonLink
  and FactionStandingsBar. OverviewScreen 394 → ~200 lines, now an orchestrator.
- FactionPageScreen: use FactionStandingsBar (drop inline bars).
- FactionHallOfFame: use shared rankMedalColorClass.

Refs #106
