+++ src/features/overview/components/OverviewScreen.tsx
@@ -1,15 +1,15 @@
 'use client';
 
-import { ArrowRight, Flame } from 'lucide-react';
-import Link from 'next/link';
 import { useLocale, useTranslations } from 'next-intl';
 
+import { SecondaryButtonLink } from '@/components/ButtonLink';
 import { factionPath } from '@/features/factions/lib/factionSlug';
-import { useClanHud } from '@/features/factions/hooks/useClanHud';
 import {
   resolveWeeklyDisplayLabel,
   useWeeklyChallenge,
 } from '@/features/overview/hooks/useWeeklyChallenge';
+import { OverviewFactionCard } from '@/features/overview/components/OverviewFactionCard';
+import { OverviewHeroCard } from '@/features/overview/components/OverviewHeroCard';
 import { EventCard } from '@/features/events/components/EventCard';
 import { useActiveEvents } from '@/features/events/hooks/useActiveEvents';
 import { ComebackWeekBanner } from '@/features/growth/components/ComebackWeekBanner';
@@ -17,39 +17,13 @@ import { WeeklyChallengeInvite } from '@/features/growth/components/WeeklyChalle
 import { useOptionalAppProfile } from '@/features/appshell/context/AppProfileContext';
 import { useProfileRating } from '@/features/profile/hooks/useProfileRating';
 import { getComebackEligibility } from '@/lib/growth/comebackWeek';
-import { getDivisionBadgeClasses } from '@/lib/divisions';
-import { getFactionBadgeClasses } from '@/lib/factionStyles';
 import { getWeeklyTimeRemaining } from '@/lib/weekly';
-import type { Faction } from '@/lib/types/database.types';
-
-function PrimaryButtonLink({ href, label }: { href: string; label: string }) {
-  return (
-    <Link
-      href={href}
-      className="inline-flex w-full items-center justify-center rounded-md bg-primary px-4 py-3 text-sm font-black uppercase tracking-[0.18em] text-primary-foreground hover:opacity-90 focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring"
-    >
-      {label}
-    </Link>
-  );
-}
-
-function SecondaryButtonLink({ href, label }: { href: string; label: string }) {
-  return (
-    <Link
-      href={href}
-      className="inline-flex w-full items-center justify-center rounded-md border border-border bg-background px-4 py-3 text-sm font-black uppercase tracking-[0.18em] text-foreground hover:bg-muted focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring"
-    >
-      {label}
-    </Link>
-  );
-}
 
 export function OverviewScreen() {
   const locale = useLocale();
   const tApp = useTranslations('app');
   const t = useTranslations('overview');
   const tWeekly = useTranslations('weekly');
-  const tFactions = useTranslations('factions');
 
   const { state: weeklyState } = useWeeklyChallenge();
   const { state: eventsState } = useActiveEvents();
@@ -59,10 +33,7 @@ export function OverviewScreen() {
   const clanHref = `/${locale}/app/clan`;
   const readyProfile = appProfile;
   const userFaction = readyProfile?.faction ?? null;
-  const factionBadge = userFaction ? getFactionBadgeClasses(userFaction) : null;
-  const factionName = userFaction ? tFactions(userFaction) : null;
   const factionHref = userFaction ? factionPath(locale, userFaction) : clanHref;
-  const divisionBadge = readyProfile ? getDivisionBadgeClasses(readyProfile.division) : null;
 
   const weekly = weeklyState.status === 'ready' ? weeklyState.weekly : null;
   const weeklyLabel = weekly ? resolveWeeklyDisplayLabel(weekly) : null;
@@ -105,80 +76,17 @@ export function OverviewScreen() {
         />
       ) : null}
 
-      {/* HERO — identity + rating + streak + the single primary action */}
       {readyProfile ? (
-        <article className="rounded-lg border border-border bg-card p-5">
-          <div className="flex items-start justify-between gap-4">
-            <div className="min-w-0">
-              <p className="truncate text-xl font-black uppercase tracking-tight">
-                {readyProfile.username ?? '—'}
-              </p>
-              <div className="mt-2 flex flex-wrap items-center gap-2">
-                {divisionBadge ? (
-                  <span
-                    className={[
-                      'inline-flex items-center rounded-sm border px-2 py-1 text-[10px] font-black uppercase tracking-[0.18em]',
-                      divisionBadge.bg,
-                      divisionBadge.text,
-                      divisionBadge.border,
-                    ].join(' ')}
-                  >
-                    {readyProfile.division}
-                  </span>
-                ) : null}
-                {userFaction && factionBadge && factionName ? (
-                  <span
-                    className={[
-                      'inline-flex items-center rounded-sm border px-2 py-1 text-[10px] font-black uppercase tracking-[0.18em]',
-                      factionBadge.bg,
-                      factionBadge.text,
-                      factionBadge.border,
-                    ].join(' ')}
-                  >
-                    {factionName}
-                  </span>
-                ) : null}
-              </div>
-            </div>
-            <div
-              className="flex shrink-0 items-center gap-2"
-              role="group"
-              aria-label={t('streakLine', { days: readyProfile.streak_days })}
-            >
-              <Flame
-                className={
-                  readyProfile.streak_days > 0
-                    ? 'h-5 w-5 shrink-0 text-accent'
-                    : 'h-5 w-5 shrink-0 text-muted-foreground'
-                }
-                aria-hidden
-              />
-              <span className="text-lg font-black tabular-nums tracking-tight">
-                {t('streakValue', { days: readyProfile.streak_days })}
-              </span>
-            </div>
-          </div>
-
-          <div className="mt-5 flex items-end justify-between gap-4 border-t border-border pt-5">
-            <div>
-              <p className="text-[10px] font-black uppercase tracking-[0.2em] text-muted-foreground">
-                {t('heroRating')}
-              </p>
-              <p className="mt-1 text-5xl font-black tabular-nums leading-none tracking-tight">
-                {ratingValue !== null ? ratingValue : '—'}
-              </p>
-              <p className="mt-2 text-xs text-muted-foreground">{t('heroRatingHint')}</p>
-            </div>
-          </div>
-
-          <div className="mt-5">
-            {showComeback ? (
-              <SecondaryButtonLink href={primaryHref} label={primaryLabel} />
-            ) : (
-              <PrimaryButtonLink href={primaryHref} label={primaryLabel} />
-            )}
-          </div>
-        </article>
+        <OverviewHeroCard
+          username={readyProfile.username}
+          division={readyProfile.division}
+          faction={userFaction}
+          streakDays={readyProfile.streak_days}
+          ratingValue={ratingValue}
+          primaryHref={primaryHref}
+          primaryLabel={primaryLabel}
+          primaryAsSecondary={showComeback}
+        />
       ) : (
         <article className="rounded-lg border border-border bg-card p-5">
           <p className="text-sm text-muted-foreground">{t('loading')}</p>
@@ -188,12 +96,7 @@ export function OverviewScreen() {
       {/* CONTENT ROW — faction standing + weekly (quiet, tappable) */}
       <div className="grid gap-4 md:grid-cols-2">
         {userFaction ? (
-          <FactionStandingCard
-            href={factionHref}
-            userFaction={userFaction}
-            t={t}
-            tFactions={tFactions}
-          />
+          <OverviewFactionCard href={factionHref} userFaction={userFaction} />
         ) : (
           <article className="rounded-lg border border-border bg-card p-5">
             <p className="text-xs font-black uppercase tracking-[0.18em] text-muted-foreground">
@@ -281,114 +184,5 @@ export function OverviewScreen() {
         ) : null}
       </article>
     </section>
-  );
-}
-
-function FactionStandingCard({
-  href,
-  userFaction,
-  t,
-  tFactions,
-}: {
-  href: string;
-  userFaction: Faction;
-  t: ReturnType<typeof useTranslations>;
-  tFactions: ReturnType<typeof useTranslations>;
-}) {
-  const { state } = useClanHud();
-
-  const rankings =
-    state.status === 'ready' ? [...state.rankings].sort((a, b) => b.points - a.points) : [];
-  const myIndex = rankings.findIndex((r) => r.faction === userFaction);
-  const myRow = myIndex >= 0 ? rankings[myIndex] : null;
-  const rank = myIndex >= 0 ? myIndex + 1 : null;
-  const leaderPoints = rankings[0]?.points ?? 0;
-  const maxPoints = Math.max(1, leaderPoints);
-  const gapText =
-    myRow && rank
-      ? rank === 1
-        ? t('factionLeadingBy', { gap: myRow.points - (rankings[1]?.points ?? 0) })
-        : t('factionGapToFirst', { gap: leaderPoints - myRow.points })
-      : null;
-
-  return (
-    <Link
-      href={href}
-      aria-label={t('viewFactionAria')}
-      className="group block rounded-lg border border-border bg-card p-5 hover:border-primary/40 focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring"
-    >
-      <div className="flex items-center justify-between gap-3">
-        <p className="text-xs font-black uppercase tracking-[0.18em] text-muted-foreground">
-          {t('factionStandingTitle')}
-        </p>
-        <ArrowRight
-          className="h-4 w-4 shrink-0 text-muted-foreground transition group-hover:translate-x-0.5 group-hover:text-foreground"
-          aria-hidden
-        />
-      </div>
-
-      {state.status === 'loading' ? (
-        <p className="mt-3 text-sm text-muted-foreground">{t('loading')}</p>
-      ) : null}
-
-      {state.status === 'error' ? (
-        <p className="mt-3 text-sm text-muted-foreground">{t('error')}</p>
-      ) : null}
-
-      {state.status === 'ready' && myRow && rank ? (
-        <div className="mt-3 space-y-4">
-          <div className="flex items-end justify-between gap-4">
-            <div>
-              <p className="text-[10px] font-black uppercase tracking-[0.2em] text-muted-foreground">
-                {t('factionRankLabel')}
-              </p>
-              <p className="mt-1 text-4xl font-black tabular-nums leading-none">#{rank}</p>
-            </div>
-            <div className="text-right">
-              <p className="text-[10px] font-black uppercase tracking-[0.2em] text-muted-foreground">
-                {t('factionPointsLabel')}
-              </p>
-              <p className="mt-1 text-2xl font-black tabular-nums leading-none text-accent">
-                {myRow.points.toLocaleString()}
-              </p>
-            </div>
-          </div>
-
-          {gapText ? (
-            <p className="text-xs font-black uppercase tracking-[0.14em] text-accent">{gapText}</p>
-          ) : null}
-
-          <div className="space-y-2">
-            {rankings.map((row) => {
-              const mine = row.faction === userFaction;
-              return (
-                <div key={row.faction} className="flex items-center gap-2">
-                  <span
-                    className={[
-                      'w-24 shrink-0 truncate text-[10px] font-black uppercase tracking-[0.14em]',
-                      mine ? 'text-foreground' : 'text-muted-foreground',
-                    ].join(' ')}
-                  >
-                    {tFactions(row.faction)}
-                  </span>
-                  <span className="h-2 flex-1 overflow-hidden rounded-full bg-muted">
-                    <span
-                      className={
-                        mine ? 'block h-full bg-accent' : 'block h-full bg-muted-foreground/50'
-                      }
-                      style={{ width: `${Math.round((row.points / maxPoints) * 100)}%` }}
-                    />
-                  </span>
-                </div>
-              );
-            })}
-          </div>
-
-          <p className="text-xs text-muted-foreground">
-            {t('factionYourShare', { points: state.myContribution.toLocaleString() })}
-          </p>
-        </div>
-      ) : null}
-    </Link>
   );
 }
